got the skeleton code for the step process UI
The dropdown menu has the correct style (as displayed in figma) and is rendered in a fixed position on the screen. The dropdown can be minimised/maximised and dragged across the screen using javascript. The dropdown items have a constrained width (to the width of the dropdown button) and a max-height. If the dropdown content exceeds a height, the menu is scrollable.