Added api call functionality for two features
These features are "Help Me" and "Hint code" and work as follows:

Help Me provides pseudo code to the user of how the student might complete the step they are stuck on.

Hint code generates the next logical line of code based on the step process and the current user code.