Am now getting a 500 internal server error
Made changes to the following:

BaseUserHandler:
Added a new method get_exercise_steps(exercise_id) to retrieve the exercise steps from the database based on the provided exercise_id

ExerciseStepsHandler:
Utilized the get_exercise_steps method from BaseUserHandler to fetch the exercise steps from the database.
This handler should now have the exercise_steps being appended to the API call made to the model

HTML file (exercise.html)
Updated the generateDropdownItems to include the extra exercise_id parameter in the AJAX request when fetching the data from the endpoint.

This code is currently encountering a 500 internal server error which is likely due to the absence of dummy/ real data in the database.

When the BaseUserHandler is attempting to fetch exercise steps from the database, it expects to recieve valid data. However, since the database does not contain any data its resulting in an internal server side error.