UI is now finished
There are loading screens whenever waiting for a response from the gpt model. There are data tooltips for each of the added features. There is also a new modal for the hint so users can check the hint before appending it to the editor. They can also regenerate the hint.